pyproject.toml
[project]
name = "gmap-screenshot-engine"
version = "0.1.0"
description = "Gmap Screenshot Engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.25",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "scrapy>=2.13,<2.14",
    "scrapy-playwright>=0.0.44",
    "scrapy-redis>=0.9.1",
    "scrapyd>=1.6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
